+++ fem/CMakeLists.txt
@@ -13,6 +13,7 @@ set(SRCS
   bilinearform.cpp
   bilinearform_ext.cpp
   bilininteg.cpp
+  bilininteg_pa.cpp
   coefficient.cpp
   datacollection.cpp
   doftoquad.cpp
@@ -23,6 +24,7 @@ set(SRCS
   fespace.cpp
   fespace_ext.cpp
   geom.cpp
+  geom_ext.cpp
   gridfunc.cpp
   hybridization.cpp
   intrules.cpp
@@ -49,6 +51,7 @@ set(HDRS
   fespace.hpp
   fespace_ext.hpp
   geom.hpp
+  geom_ext.hpp
   gridfunc.hpp
   hybridization.hpp
   intrules.hpp
