+++ fem/bilininteg.hpp
@@ -29,8 +29,7 @@ public:
    AbstractBilinearFormIntegrator(const IntegrationRule *ir = NULL) :
       NonlinearFormIntegrator(ir) { }
    virtual ~AbstractBilinearFormIntegrator() { };
-   virtual void Setup(const FiniteElementSpace*, const IntegrationRule*) { }
-   virtual void Assemble() { }
+   virtual void Assemble(const FiniteElementSpace&) { }
    virtual void MultAdd(Vector&, Vector&) { }
    virtual void MultTransposeAdd(Vector&, Vector&) { }
    virtual void AssembleElementMatrix(const FiniteElement&,
@@ -47,8 +46,7 @@ public:
    BilinearPAFormIntegrator(const IntegrationRule *ir = NULL) :
       AbstractBilinearFormIntegrator(ir) { }
    virtual ~BilinearPAFormIntegrator() { };
-   virtual void Setup(const FiniteElementSpace*, const IntegrationRule*) =0;
-   virtual void Assemble() =0;
+   virtual void Assemble(const FiniteElementSpace&) =0;
    virtual void MultAdd(Vector&, Vector&) =0;
 };
 
@@ -58,16 +56,15 @@ public:
 class PAMassIntegrator: public BilinearPAFormIntegrator
 {
 private:
-   Vector op;
+   Coefficient *Q;
+   Vector vec;
    DofToQuad *maps;
-   const FiniteElementSpace *fes;
-   const IntegrationRule *ir;
+   int dim, ne, nq, dofs1D, quad1D;
 public:
-   PAMassIntegrator () : BilinearPAFormIntegrator(), op(), maps(NULL) { }
+   PAMassIntegrator (Coefficient &q, const IntegrationRule *ir = NULL) :
+      BilinearPAFormIntegrator(ir), Q(&q), vec(), maps(NULL) { }
    ~PAMassIntegrator() { delete maps; }
-   void Setup(const FiniteElementSpace*, const IntegrationRule*);
-   void Assemble();
-   void SetOperator(Vector&);
+   void Assemble(const FiniteElementSpace&);
    void MultAdd(Vector&, Vector&);
 };
 
@@ -78,16 +75,14 @@ class PADiffusionIntegrator: public BilinearPAFormIntegrator
 {
 private:
    Coefficient *Q;
-   Vector op;
+   Vector vec;
    DofToQuad *maps;
-   const FiniteElementSpace *fes;
-   const IntegrationRule *ir;
+   int dim, ne, nq, dofs1D, quad1D;
 public:
-   PADiffusionIntegrator (Coefficient &q) :
-      BilinearPAFormIntegrator(), Q(&q), op(), maps(NULL) { }
+   PADiffusionIntegrator (Coefficient &q, const IntegrationRule *ir = NULL) :
+      BilinearPAFormIntegrator(ir), Q(&q), vec(), maps(NULL) { }
    ~PADiffusionIntegrator() { delete maps; }
-   void Setup(const FiniteElementSpace*, const IntegrationRule*);
-   void Assemble();
+   void Assemble(const FiniteElementSpace&);
    void MultAdd(Vector&, Vector&);
 };
